requirements.txt
# Core
requests==2.31.0
python-dotenv==1.0.0

# Google APIs
google-generativeai==0.7.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0

# Trends Research
pytrends==4.9.2

# TTS Voice Generation
edge-tts==6.1.9

# Image & Video Processing
Pillow==10.3.0
requests==2.31.0

# Audio Analysis (Quality Gate)
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4

# Data
sqlite3  # stdlib

# Utilities
asyncio  # stdlib
uuid     # stdlib
